reformatting to minter contract
